simple_lmk: Implemented a simple adaptiveness mechanism.
This is a simple mechanism which extends SimpleLMK to support a kind of "adaptive aggressiveness".

By utilizing pre-defined memory pressure levels, it calculates how much memory to free. This allows SimpleLMK to be aggressive on-demand without affecting UX that much.

Be careful to not set CONFIG_ANDROID_SIMPLE_LMK_MINFREE too high or too low as that would break the logic of the mechanism.